four_seasons_parser update
